requirements.txt
torch>=2.4
torchvision>=0.19
numpy>=1.24
Pillow>=10.0
opencv-python-headless>=4.10
hydra-core>=1.3
sam2
tqdm>=4.66

